+++ database/migrations/2025_11_25_022219_create_keranjangs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('produk_id')->constrained('produks')->onDelete('cascade');
-            $table->integer('jumlah');
+            $table->unsignedBigInteger('jumlah');
             $table->timestamps();
         });
     }
